"small performance increase" (read desc for funny)
When we update a tilemap at runtime, we query an update to the spritebatch which redoes the entirety of the spritebatch, which requires a recompiling of all the tileset pieces, which is a huge performance bottleneck.

It turns out I forgot to set the Query back to false after we do it. So this entire time, we've been redoing the spritebatch EVERY frame after the first redraw.